Create roll dice method

+++ week5/day1/src/GameCharacter.java
@@ -1,9 +1,9 @@
 import java.awt.image.BufferedImage;
 
 public class GameCharacter extends GameObject{
-  double hp;
-  double dp;
-  double sp;
+  int hp;
+  int dp;
+  int sp;
 
   public GameCharacter(BufferedImage image, int posX, int posY) {
     super(image, posX, posY);
